requirements.txt
numpy==1.21.0
pyenchant==3.2.2
